+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -14077,6 +14077,13 @@ $DGURU_Completers = {
             break
         }
 
+        # Amazon.DevOpsGuru.OptInStatus
+        "Update-DGURUServiceIntegration/ServiceIntegration_LogsAnomalyDetection_OptInStatus"
+        {
+            $v = "DISABLED","ENABLED"
+            break
+        }
+
         # Amazon.DevOpsGuru.OrganizationResourceCollectionType
         "Get-DGURUOrganizationResourceCollectionHealth/OrganizationResourceCollectionType"
         {
@@ -14091,6 +14098,13 @@ $DGURU_Completers = {
         }
         {
             $v = "AWS_CLOUD_FORMATION","AWS_SERVICE","AWS_TAGS"
+            break
+        }
+
+        # Amazon.DevOpsGuru.ResourcePermission
+        "Get-DGURUMonitoredResourceList/Filters_ResourcePermission"
+        {
+            $v = "FULL_PERMISSION","MISSING_PERMISSION"
             break
         }
 
@@ -14114,9 +14128,11 @@ $DGURU_map = @{
     "EventSources_AmazonCodeGuruProfiler_Status"=@("Update-DGURUEventSourcesConfig")
     "Filters_DataSource"=@("Get-DGURUEventList")
     "Filters_EventClass"=@("Get-DGURUEventList")
+    "Filters_ResourcePermission"=@("Get-DGURUMonitoredResourceList")
     "Locale"=@("Get-DGURURecommendationList")
     "OrganizationResourceCollectionType"=@("Get-DGURUOrganizationResourceCollectionHealth")
     "ResourceCollectionType"=@("Get-DGURUResourceCollection","Get-DGURUResourceCollectionHealth")
+    "ServiceIntegration_LogsAnomalyDetection_OptInStatus"=@("Update-DGURUServiceIntegration")
     "StatusFilter_Any_Type"=@("Get-DGURUOrganizationInsightList")
     "StatusFilter_Closed_Type"=@("Get-DGURUOrganizationInsightList")
     "StatusFilter_Ongoing_Type"=@("Get-DGURUOrganizationInsightList")
@@ -14189,8 +14205,10 @@ $DGURU_SelectMap = @{
                "Get-DGURUCostEstimation",
                "Get-DGURUResourceCollection",
                "Get-DGURUAnomaliesForInsightList",
+               "Get-DGURUAnomalousLogGroupList",
                "Get-DGURUEventList",
                "Get-DGURUInsightList",
+               "Get-DGURUMonitoredResourceList",
                "Get-DGURUNotificationChannelList",
                "Get-DGURUOrganizationInsightList",
                "Get-DGURURecommendationList",
@@ -44932,7 +44950,8 @@ $SME_SelectCompleters = {
 }
 
 $SME_SelectMap = @{
-    "Select"=@("Get-SMEDeviceRegistration",
+    "Select"=@("Get-SMEDeployment",
+               "Get-SMEDeviceRegistration",
                "Send-SMEHeartbeat")
 }