+++ modules/AWSPowerShell/Cmdlets/Glue/Basic/Start-GLUEJobRun-Cmdlet.cs
@@ -141,8 +141,8 @@ namespace Amazon.PowerShell.Cmdlets.GLUE
         /// <para>
         /// <para>The <code>JobRun</code> timeout in minutes. This is the maximum time that a job run
         /// can consume resources before it is terminated and enters <code>TIMEOUT</code> status.
-        /// The default is 2,880 minutes (48 hours). This overrides the timeout value set in the
-        /// parent job.</para>
+        /// This value overrides the timeout value set in the parent job.</para><para>Streaming jobs do not have a timeout. The default for non-streaming jobs is 2,880
+        /// minutes (48 hours).</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]